feat: now starting queue

+++ src/domain/file/usecases/delete-file-by-reference-use-case.ts
@@ -1,20 +1,27 @@
-import { StorageInterface } from '../../infra/storage/storage';
+import { NotAllowedError } from '../../../core/errors/not-allowed-error';
+import { ResourceNotFoundError } from '../../../core/errors/resource-not-found';
+import { StorageInterface } from '../../../infra/storage/storage';
 import { FilesRepository } from '../repositories/files-repository';
 
 interface DeleteFileByReferenceUseCaseRequest {
   file_reference: string;
+  userId: string;
 }
 
 export class DeleteFileByReferenceUseCase {
   constructor(private filesRepository: FilesRepository, private storage: StorageInterface) {}
 
   execute = async (data: DeleteFileByReferenceUseCaseRequest) => {
-    const { file_reference } = data;
+    const { file_reference, userId } = data;
+
+    if (file_reference.split('-')[0] !== userId) {
+      throw new NotAllowedError();
+    }
 
     const file = await this.filesRepository.findByReference(file_reference);
 
     if (!file) {
-      throw new Error('File not found');
+      throw new ResourceNotFoundError();
     }
 
     this.storage.delete(file_reference);
